sched_entity: reorder BORE-related member variables child_burst_cnt: type u16 -> u32

+++ include/linux/sched.h
@@ -562,7 +562,7 @@ struct sched_entity {
 	u8				burst_penalty;
 	u8				burst_score;
 	u8				child_burst;
-	u16				child_burst_cnt;
+	u32				child_burst_cnt;
 	u64				child_burst_last_cached;
 #endif // CONFIG_SCHED_BORE
